[frontend] Add SocketProvider to re-connect on auth status change
